fix: makes the homepage a reusable component
